AutostartDb-generic.cpp: missing cast.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@63397 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/generic/AutostartDb-generic.cpp
@@ -83,7 +83,7 @@ int AutostartDb::autostartModifyDb(bool fAutostart, bool fAddVM)
                     /* Check if the file was just created. */
                     if (cbFile)
                     {
-                        rc = RTFileRead(hAutostartFile, abBuf, cbFile, NULL);
+                        rc = RTFileRead(hAutostartFile, abBuf, (size_t)cbFile, NULL);
                         if (RT_SUCCESS(rc))
                         {
                             rc = RTStrToUInt32Ex(abBuf, NULL, 10 /* uBase */, &cAutostartVms);
